BE/requirements.txt
# Requirements for EduGrade AI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
firebase-admin>=6.4.0
google-genai>=0.2.0
openai>=1.12.0
httpx>=0.26.0
pydantic>=2.0.0
python-multipart
